+++ src/app/armor/components/customSortBy/FormulaField.tsx
@@ -13,8 +13,21 @@ export default function FormulaField(props: {
         props.sortBy
     );
     const TAB = "    ";
-
     let tabs: number = 0;
+
+    /**
+     * Maps a token group to a string where each token is on a new line
+     * with appropriate indentation based on the number of tabs.
+     * If the token group contains only one token and the token includes
+     * an open parenthesis, it will indent the token and increment the
+     * number of tabs. If the token group contains only one token and the
+     * token is a close parenthesis, it will unindent the token and
+     * decrement the number of tabs. If the token group contains more than
+     * one token, it will recursively call itself on each token group in the
+     * child token groups and join the results together.
+     * @param tokenGroup - a string array of tokens
+     * @returns a string where each token is on a new line with appropriate indentation
+     */
     function mapTokenGroupToLines(tokenGroup: string[]): string {
         // if length is 1, process the token
         if (tokenGroup.length === 1) {