+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.11.7] - 2025-11-17
+
+### Added
+
+-   Invert option for custom armor sorting
+
+### Fixed
+
+-   Typos
+-   Bug in final stat calculation on Class page
+
 ## [2.11.6] - 2025-06-18
 
 ### Fixed